refactor: move relation info to edge.ts
